requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
requests>=2.31.0
schedule>=1.2.0
